update-restructure-types-and-redux
