+++ src/components/TrialDetail/DetailBoard/TrialInfo/Challenges/index.tsx
@@ -1,4 +1,4 @@
-import type { TrialDetailSupa } from "@/components/types/TrialDetailSupa";
+import type { TrialDetailSupa } from "@/types/TrialDetailSupa";
 
 interface acceptProps {
   trial: TrialDetailSupa[];
