+++ src/components/TrialDetail/DetailBoard/UploadCalender/DayBox.tsx
@@ -1,6 +1,6 @@
 import { useRef, useEffect, useState } from "react";
 import { useSelector } from "react-redux";
-import { RootState } from "@/state/store";
+import { RootState } from "@/store";
 import PopupCard from "./PopupCard";
 
 interface acceptProps {
